[fish]: Add `confirm` function inspired by `click.confirm`
This function can be used to ask for user confirmation, for example a
function can ask for user confirmation to delete a file.
```fish
if (confirm "Are you sure you want to delete this file?")
  rm file
end
```

Default value of `confirm` is false that is if the prompt is left blank
function will return with status code of 1. This behavior can be changed
with `-d/--default` option.

Depending on the default value `[y/N]` or `[Y/n]` is appended to the
prompt with suffix of ': '. Suffix can be changed with `-s/--suffix`
option.